+++ src/main/java/me/marin/lockout/server/LockoutServer.java
@@ -389,6 +389,11 @@ public class LockoutServer {
 
         world.setTimeOfDay(0);
 
+        // Unfreeze ticks when lockout/blackout game starts
+        var unfreezeCommand = "tick unfreeze";
+        var unfreezeParseResults = server.getCommandManager().getDispatcher().parse(unfreezeCommand, server.getCommandSource());
+        server.getCommandManager().execute(unfreezeParseResults, unfreezeCommand);
+
         for (int i = 3; i >= 0; i--) {
             if (i > 0) {
                 final int secs = i;