+++ src/main/java/me/marin/lockout/server/handlers/ServerStartedEventHandler.java
@@ -78,6 +78,11 @@ public class ServerStartedEventHandler implements ServerLifecycleEvents.ServerSt
             }
             long end = System.currentTimeMillis();
             Lockout.log("Located " + BIOME_LOCATE_DATA.size() + " biomes and " + STRUCTURE_LOCATE_DATA.size() + " structures in " + String.format("%.2f", ((end-start)/1000.0)) + "s!");
+            
+            // Freeze ticks until lockout/blackout game starts
+            var freezeCommand = "tick freeze";
+            var parseResults = server.getCommandManager().getDispatcher().parse(freezeCommand, server.getCommandSource());
+            server.getCommandManager().execute(parseResults, freezeCommand);
         });
     }
 }